styles: made style changes in the employee profile, and changed the naming convention
