+++ src/components/Layout/Layout.module.css
@@ -1,6 +1,6 @@
 .container {
   width: 100%;
   height: 100vh;
-  padding: 76px 16px 16px 286px;
+  padding: 80px 16px 16px 286px;
   background: var(--bg-white);
 }
