+++ src/feature/EmployeeProfile/EmployeeProfile.module.css
@@ -4,7 +4,6 @@
   padding: 10px;
   box-shadow: var(--box-shadow);
   transition: box-shadow 0.3s;
-  margin-right: 20px;
 }
 
 .employee__profile__section1:hover {
@@ -30,15 +29,13 @@
 
 .vertical__line {
   border-left: 3px dotted var(--primary-black-4);
-  margin-top: 25px;
-  margin-left: 5px;
+  margin-right: 15px;
   height: 120px;
 }
 
 .information__sections {
   display: flex;
-  align-items: flex-start;
-  width: 81vw;
+  
 }
 
 .vertical__contact {
