+++ src/feature/EmployeeProfile/component/PassportInfo/PassportInfo.module.css
@@ -1,4 +1,4 @@
-.emergencyContactContainer {
+.passportInfoContainer {
   margin-top: 20px;
   margin-right: 20px;
   background-color: var(--bg-color);
@@ -7,16 +7,16 @@
   padding: 20px;
   box-shadow: var(--box-shadow);
   transition: box-shadow 0.3s;
-  padding-bottom: 25px;
   position: relative;
 }
 
-.emergencyContactContainer:hover {
+.passportInfoContainer:hover {
   box-shadow: var(--box-shadow-hover);
 }
 
 .infoEdit {
   display: flex;
+  flex-direction: row;
   justify-content: space-between;
 }
 
@@ -24,16 +24,6 @@
   margin-bottom: 20px;
   font-weight: var(--font-bold);
   font-size: 20px;
-}
-
-.editIcon {
-  color: grey;
-  cursor: pointer;
-}
-
-.infoColumn {
-  display: flex;
-  flex-direction: column;
 }
 
 .infoRow {
@@ -47,18 +37,6 @@
   min-width: 250px;
 }
 
-.infoValue {
-  margin-left: 10px;
-  color: var(--primary-black-4);
-}
-
-.thinLine {
-  height: 1px;
-  width: 100%;
-  background-color: #ccc;
-  margin: 10px 0;
-}
-
 .contact__info__right {
   top: 0;
   right: 0;
@@ -70,5 +48,10 @@
   display: flex;
   align-items: flex-start;
   cursor: pointer;
-  color: var(--primary-black-4)
+  color: var(--primary-black-4);
+}
+
+.infoValue {
+  color: var(--primary-black-4);
+  margin-left: 10px;
 }
